combine `Any` and `misc` files into `misc`
