+++ src/utilityFunctions/__tests__/misc.spec.ts
@@ -1,5 +1,14 @@
 import isCI from 'is-ci'
-import { cast, exactly, failCI, unsafeCast } from '../misc'
+import {
+    cast,
+    entries,
+    exactly,
+    failCI,
+    hasPropertyPredicate,
+    isNullOrUndefined,
+    runUntil,
+    unsafeCast,
+} from '../misc'
 import { PowerAssert } from 'typed-nodejs-assert'
 import { NonNullish } from '../../utilityTypes/misc'
 // eslint-disable-next-line @typescript-eslint/no-var-requires,@typescript-eslint/no-unsafe-assignment -- https://github.com/detachHead/typed-nodejs-assert#with-power-assert
@@ -479,4 +488,79 @@ test('failCI', () => {
     console.log({ isCI })
     if (isCI) assert.throws(failCI)
     else assert.doesNotThrow(failCI)
+})
+
+test('entries', () => {
+    exactly(
+        [
+            ['foo', 1],
+            ['bar', 'baz'],
+        ] as ['foo' | 'bar', string | number][],
+        entries({ foo: 1, bar: 'baz' }),
+    )
+})
+
+describe('hasPropertyPredicate', () => {
+    const value = {} as unknown
+    describe('inferred', () => {
+        test('one value', () => {
+            if (hasPropertyPredicate(value, 'foo')) exactly<{ foo: unknown }>()(value)
+        })
+        test('union', () => {
+            if (hasPropertyPredicate(value, 'length' as 'length' | 'asdf'))
+                exactly<NonNullish>()(value)
+        })
+        test('not known at compiletime', () => {
+            if (hasPropertyPredicate(value, 'length' as string)) exactly<unknown>()(value)
+        })
+    })
+    test('explicit', () => {
+        if (hasPropertyPredicate<unknown[]>(value, 'length')) exactly<unknown[]>()(value)
+    })
+    describe('undefined/null', () => {
+        test('undefined', () => {
+            hasPropertyPredicate(undefined, 'length')
+        })
+        test('null', () => {
+            hasPropertyPredicate(null, 'length')
+        })
+    })
+})
+
+describe('runUntil', () => {
+    test('rejects', async () => {
+        let isDone = false
+        setTimeout(() => (isDone = true), 1000)
+        await expect(
+            runUntil(
+                () => new Promise<boolean>((res) => setTimeout(() => res(isDone), 10)),
+                100,
+            ),
+        ).rejects.toThrow("runUntil failed because the predicate didn't return true in 100 ms")
+    })
+    test('resolves', async () => {
+        let isDone = false
+        setTimeout(() => (isDone = true), 80)
+        await runUntil(
+            () => new Promise<boolean>((res) => setTimeout(() => res(isDone), 10)),
+            100,
+        )
+    })
+})
+
+describe('isNullOrUndefined', () => {
+    describe('true', () => {
+        test('null', () => {
+            exactly(true, isNullOrUndefined(null))
+        })
+        test('undefined', () => {
+            exactly(true, isNullOrUndefined(undefined))
+        })
+    })
+    test('false', () => {
+        exactly(false, isNullOrUndefined('foo'))
+    })
+    test('not known at compiletime', () => {
+        exactly<boolean>()(isNullOrUndefined(1 as unknown))
+    })
 })
