--- src/utilityTypes/Any.ts
@@ -1,11 +1,0 @@
-import { Keys as TsToolbeltKeys } from 'ts-toolbelt/out/Any/Keys'
-
-/**
- * like ts-toolbelt's `Keys` except it doesn't include number (for objects where you know all of the keys)
- *
- * in some cases, `ts-toolbelt`'s `KnownKeys` will probably work for you, but it doesn't seem to work properly for arrays
- */
-export type Keys<T> = Exclude<TsToolbeltKeys<T>, number>
-
-/** compiletime version of {@link ObjectConstructor.entries} */
-export type Entries<T> = [Keys<T>, T[Keys<T>]][]
