+++ src/utilityTypes/String.ts
@@ -4,7 +4,7 @@ import { Length } from 'ts-toolbelt/out/String/Length'
 import { AnyKey, IsNever } from 'tsdef'
 import { Cast } from 'ts-toolbelt/out/Any/Cast'
 import { IndexOfLongestString, TupleOf } from './Array'
-import { Keys } from './Any'
+import { Keys } from './misc'
 import { ListOf } from 'ts-toolbelt/out/Union/ListOf'
 import { Head as ArrayHead } from 'ts-toolbelt/out/List/Head'
 import { Tail as ArrayTail } from 'ts-toolbelt/out/List/Tail'
